+++ IdleEngine/ICollidable.cs
@@ -19,7 +19,7 @@ namespace IdleEngine
     {
         public CollisionType CollisionType { get; protected set; }
         public Vector2 Position { get; protected set; }
-        public int Radius { get; protected set; }
+        public float Radius { get; protected set; }
         public Rectangle Bounds { get; protected set; }
         public bool IsCollidable { get; protected set; }
         public Vector2 Origin { get; set; }
